updates useres to include loginUser
